Fix the set method for mono1 (prepare)
The previous implementation from #3 worked fine only for the first
run. On the second run the old_value for some reason was set to 1,
and the (new) value was set to 0 on the second run, seems like the
values are kept from the previous successful run. That causes the
`RE(bps.mv(mono1, 'prepare'))` to finish before the actual
preparation of the trajectory is done
(`XF:07BMA-OP{MC:03}TrajInitPlc-Read` was set back to `0` from `1`)

Here is the log of values what was happening on the first (In [2])
and second (In [4]) runs:

====================================
In [1]: %mov mono1.energy 20_000

In [2]: RE(bps.mv(mono1, 'prepare'))
=== old value: 0.0, <class 'float'>
=== new value: 0.0, <class 'float'>

=== old value: 0.0, <class 'float'>. No progress bar available.]
=== new value: 1.0, <class 'float'>

=== old value: 1.0, <class 'float'>
=== new value: 0.0, <class 'float'>

Out[2]: ()

In [3]: %mov mono1.energy 20_000

In [4]: RE(bps.mv(mono1, 'prepare'))
=== old value: 1.0, <class 'float'>
=== new value: 0.0, <class 'float'>
====================================